fix: Move popups to modals folder
